refactor(memory): register operations from catalog
